Add function to handle HTTP method definitions
While defining a new route for your web service you must specify the
HTTP method, the URL and the HTTP handler. The URL can be either static
or dynamic like "/lorem/ipsum" and "/lorem/ipsum/:id/:name" respectively
you will notice that in the latter case the server will pass all the
dynamic named parameters from the URL to the HTTP handler via the solid
context package, so in this case you will be able to access the data
using this code:

ctx := r.Context()
id := ctx.Value("id").(int)
name := ctx.Value("name").(string)

The HTTP handler is compatible with the standard http.HandlerFunc so you
can attach more middlewares to improve the cache, the response time and
a new format for the logger, or any other assistance for your project.