Add dispatcher to determine which handler to execute
If the defined URL contains dynamic parameters we need to check if the
requested URL is longer than the defined URL without the dynamic
sections, this means that the requested URL must be longer than the
clean defined URL.

- Defined (Raw):   /lorem/ipsum/dolor/:unique
- Defined (Clean): /lorem/ipsum/dolor
- Req. URL (Bad):  /lorem/ipsum/dolor
- Req. URL (Semi): /lorem/ipsum/dolor/
- Req. URL (Good): /lorem/ipsum/dolor/something

Notice how the good requested URL has more characters than the clean
defined URL, the extra characters will be extracted and converted into
variables to be passed to the handler. The bad requested URL matches the
exact same clean defined URL but has no extra characters, so variable
"unique" will be empty which is non-processable. The semi good requested
URL contains one character more than the clean defined URL, the extra
character is simply a forward slash, which means the dynamic variable
"unique" will be empty but at least it was on purpose.

The requested URL must contains the same characters than the clean
defined URL, at least from index zero, the rest of the requested URL can
be different. This is to prevent malicious requests with semi valid URLs
with different roots which might translate to handlers processing
unrelated requests.