Fix additional middleware support to work late
When the HTTP handlers are registered before the additional middlewares
are added, the request is not passed through them because currently the
chain is built during the registration. This commit changes this to
allow the server to register the HTTP handlers at any time, and the
chain of HTTP handlers will run late, when the request is handled.